breaking: Change NetworkIdentity.netId from uint to ushort
- adds `static readonly Stack<ushort>` to `NetworkIdentity`
- `GetNextNetworkId` does Pop to reuse if available
- `OnDestroy` Pushes to the Stack
- Added entry to `NetworkInformationPreview` to show the `Stack.Count`
- all unit tests updated

Play tested with Room example and Multiple Additive Scenes example since those examples destroy and spawn prize objects rapidly, alongside having clients drop and rejoin with prizes respawned in between.

Also play tested with Tanks example to watch the projectile id's get added to the stack when destroyed and reuse correctly by watching the Network Information panel update as expected.

Would there need to be any changes to Weaver?

+++ Assets/Mirror/Runtime/Messages.cs
@@ -54,7 +54,7 @@ namespace Mirror
     #region System Messages requried for code gen path
     public struct CommandMessage : NetworkMessage
     {
-        public uint netId;
+        public ushort netId;
         public int componentIndex;
         public int functionHash;
         // the parameters for the Cmd function
@@ -64,7 +64,7 @@ namespace Mirror
 
     public struct RpcMessage : NetworkMessage
     {
-        public uint netId;
+        public ushort netId;
         public int componentIndex;
         public int functionHash;
         // the parameters for the Cmd function
@@ -79,7 +79,7 @@ namespace Mirror
         /// <summary>
         /// netId of new or existing object
         /// </summary>
-        public uint netId;
+        public ushort netId;
         /// <summary>
         /// Is the spawning object the local player. Sets ClientScene.localPlayer
         /// </summary>
@@ -122,17 +122,17 @@ namespace Mirror
 
     public struct ObjectDestroyMessage : NetworkMessage
     {
-        public uint netId;
+        public ushort netId;
     }
 
     public struct ObjectHideMessage : NetworkMessage
     {
-        public uint netId;
+        public ushort netId;
     }
 
     public struct UpdateVarsMessage : NetworkMessage
     {
-        public uint netId;
+        public ushort netId;
         // the serialized component data
         // -> ArraySegment to avoid unnecessary allocations
         public ArraySegment<byte> payload;
